refactor; add gcp, aws cluster spec implementations
